refactor: extract photo upload into reusable hook
