+++ src/hooks/domain/onboarding/usePhotoInput.ts
@@ -1,68 +1,21 @@
-import React from 'react';
-import { useRef, useState  } from 'react';
-import { getWebcamStream } from '@/src/utils/getWebcam';
+import { useCameraCapture } from './useCameraCapture';
+import { useImageFileInput } from './useImageFileInput';
 
 export const usePhotoInput = () => {
-	const [file, setFile] = useState<File | null>(null);
-	const [previewUrl, setPreviewUrl] = useState<string | null>(null);
-	const [isCamera, setIsCamera] = useState(false);
+	const image = useImageFileInput();
+	const camera = useCameraCapture();
 
-	const videoRef = useRef<HTMLVideoElement | null>(null);
-	const canvasRef = useRef<HTMLCanvasElement | null>(null);
-	const fileInputRef = useRef<HTMLInputElement | null>(null);
+	const captureFromCamera = () => {
+		const dataUrl = camera.capturePhoto();
+		if (!dataUrl) return;
 
-	/* ---------- gallery ---------- */
-	const openFilePicker = () => {
-		fileInputRef.current?.click();
-	};
-
-	const handleChangeFile = (e: React.ChangeEvent<HTMLInputElement>) => {
-		const file = e.target.files?.[0];
-		if (!file) return;
-		
-		setFile(file);
-
-		const previewUrl = URL.createObjectURL(file);
-		setPreviewUrl(previewUrl);
-		e.target.value = '';
-	};
-
-	/* ---------- camera ---------- */
-	const openCamera = async () => {
-		if (!videoRef.current) return;
-
-		const stream = await getWebcamStream();
-		videoRef.current.srcObject = stream;
-		setIsCamera(true);
-	};
-
-	const capturePhoto = () => {
-		if (!videoRef.current || !canvasRef.current) return;
-
-		const canvas = canvasRef.current;
-		const video = videoRef.current;
-		const ctx = canvas.getContext('2d');
-		if (!ctx) return;
-
-		canvas.width = video.videoWidth;
-		canvas.height = video.videoHeight;
-		ctx.drawImage(video, 0, 0);
-
-		setPreviewUrl(canvas.toDataURL('image/png'));
-		setIsCamera(false);
+		image.setPreviewUrl(dataUrl);
+		image.setFile(null); // 카메라는 File 없음
 	};
 
 	return {
-		file,
-		previewUrl,
-		isCamera,
-		videoRef,
-		canvasRef,
-		fileInputRef,
-		setFile,
-		openCamera,
-		openFilePicker,
-		handleChangeFile,
-		capturePhoto,
+		...image,
+		...camera,
+		captureFromCamera,
 	};
 };
